@@ -1,17 +1,16 @@
 module github.com/tonymet/dnstrace
 
-go 1.24.0
+go 1.25.0
 
 require (
+	codeberg.org/miekg/dns v0.5.20
 	github.com/HdrHistogram/hdrhistogram-go v1.1.2
-	github.com/miekg/dns v1.1.68
 )
 
 require (
+	github.com/google/go-cmp v0.6.0 // indirect
 	github.com/stretchr/testify v1.9.0 // indirect
-	golang.org/x/mod v0.29.0 // indirect
+	golang.org/x/crypto v0.43.0 // indirect
 	golang.org/x/net v0.46.0 // indirect
-	golang.org/x/sync v0.17.0 // indirect
 	golang.org/x/sys v0.37.0 // indirect
-	golang.org/x/tools v0.38.0 // indirect
 )
