@@ -2,76 +2,53 @@ package main
 
 import (
 	"context"
-	"fmt"
+	"io"
 	"log"
+	"net"
 	"testing"
 
-	"github.com/miekg/dns"
+	"codeberg.org/miekg/dns"
 )
 
-const serverAddr = "127.0.0.1:4453"
+const serverAddr = "127.0.0.1:4454"
 
 func init() {
-	// Start a primitive DNS server for testing
-	go startTestDNSServer()
+	startTestDNSServer()
 }
 
 func startTestDNSServer() {
-	// Listen on UDP port 5353
-	udpServer := &dns.Server{Addr: serverAddr, Net: "udp"}
 	dns.HandleFunc(".", handleDNSRequest)
+	udpServer := &dns.Server{Addr: serverAddr, Net: "udp", ReusePort: true, MaxTCPQueries: -1}
 	go func() {
-		err := udpServer.ListenAndServe()
-		if err != nil {
-			log.Printf("Failed to start UDP DNS server: %s\n", err.Error())
+		if err := udpServer.ListenAndServe(); err != nil {
+			log.Printf("Failed to setup the server: %s", err.Error())
 		}
 	}()
-
-	// // Listen on TCP port 5353
-	// tcpServer := &dns.Server{Addr: serverAddr, Net: "tcp"}
-	// go func() {
-	// 	err := tcpServer.ListenAndServe()
-	// 	if err != nil {
-	// 		log.Printf("Failed to start TCP DNS server: %s\n", err.Error())
-	// 	}
-	// }()
-
-	log.Println("Test DNS server started on " + serverAddr + " (UDP/TCP)")
+	log.Println("Test DNS server started on " + udpServer.Addr + " (UDP)")
 }
 
-func handleDNSRequest(w dns.ResponseWriter, r *dns.Msg) {
+func handleDNSRequest(ctx context.Context, w dns.ResponseWriter, r *dns.Msg) {
 	m := new(dns.Msg)
-	m.SetReply(r)
-	m.Compress = false
-
+	r.Response = true
 	if r.Opcode == dns.OpcodeQuery {
 		for _, q := range r.Question {
-			switch q.Qtype {
+			switch dns.RRToType(q) {
 			case dns.TypeA:
-				if q.Name == dns.Fqdn("www.ucla.edu") {
-					rr, err := dns.NewRR(fmt.Sprintf("%s A 192.168.1.1", q.Name))
-					if err == nil {
-						m.Answer = append(m.Answer, rr)
-					}
+				if q.Header().Name == "www.ucla.edu." {
+					rr := &dns.A{Hdr: dns.Header{Name: "www.ucla.edu.", Class: dns.ClassINET},
+						A: net.IPv4zero}
+					m.Answer = append(m.Answer, rr)
 				}
 			case dns.TypeAAAA:
-				if q.Name == dns.Fqdn("www.ucla.edu") { // Example for AAAA
-					rr, err := dns.NewRR(fmt.Sprintf("%s AAAA ::1", q.Name))
-					if err == nil {
-						m.Answer = append(m.Answer, rr)
-					}
-				}
-			case dns.TypeTXT:
-				if q.Name == dns.Fqdn("www.ucla.edu") { // Example for TXT
-					rr, err := dns.NewRR(fmt.Sprintf("%s TXT \"hello world\"", q.Name))
-					if err == nil {
-						m.Answer = append(m.Answer, rr)
-					}
+				if q.Header().Name == "www.ucla.edu." { // Example for AAAA
+					rr := &dns.A{Hdr: dns.Header{Name: "www.ucla.edu.", Class: dns.ClassINET},
+						A: net.IPv6zero}
+					m.Answer = append(m.Answer, rr)
 				}
 			}
 		}
 	}
-	_ = w.WriteMsg(m)
+	_, _ = io.Copy(w, m)
 }
 
 func Test_do(t *testing.T) {
@@ -90,7 +67,8 @@ func Test_do(t *testing.T) {
 			pConcurrency = &tt.pConcurrency
 			*pServer = serverAddr
 			pQueries = []string{tt.query}
-			do(context.Background())
+			//do(context.Background())
+			do(t.Context())
 			got := countResults()
 			if tt.wantCount != got {
 				t.Errorf("got = %v, want %v", got, tt.wantCount)
@@ -112,7 +90,8 @@ func Benchmark_do(b *testing.B) {
 	for range b.N {
 		pCount = &tt.pCount
 		pConcurrency = &tt.pConcurrency
-		*pServer = serverAddr
+		//*pServer = serverAddr
+		*pServer = "127.0.0.1:5353"
 		pQueries = []string{tt.query}
 		do(context.Background())
 	}
